+++ demisto_sdk/commands/validate/tests/validators_test.py
@@ -7,7 +7,8 @@ from typing import Any, Type, Union
 
 import demisto_sdk.commands.validate.validate_manager
 import pytest
-from demisto_sdk.commands.common.constants import CONF_PATH, TEST_PLAYBOOK
+from demisto_sdk.commands.common.constants import (CONF_PATH, TEST_PLAYBOOK,
+                                                   FileType)
 from demisto_sdk.commands.common.git_tools import git_path
 from demisto_sdk.commands.common.hook_validations.base_validator import \
     BaseValidator
@@ -1014,32 +1015,56 @@ class TestValidators:
         assert validate_manager.validate_release_notes(file_path, {file_path}, modified_files, None, True) is False
 
     def test_staged(self, mocker):
+        """
+        Given
+            - staged = True flag
+            - diff on yml file
+        When
+            - Run the validate command.
+        Then
+            - Validate that only one git diff command runs and it's a staged command
+        """
         def run_command_effect(arg):
-            if arg == 'git diff --name-status --staged master':
-                return 'M\tPacks/HelloWorld/Integrations/HelloWorld.py'
+            assert arg == 'git diff --name-status --staged master'
             return 'M\tPacks/HelloWorld/Integrations/HelloWorld.yml'
 
         def is_file_yml(arg):
             assert arg == 'Packs/HelloWorld/Integrations/HelloWorld.yml'
+            return True
 
         mocker.patch('demisto_sdk.commands.validate.validate_manager.run_command', side_effect=run_command_effect)
         mocker.patch('demisto_sdk.commands.validate.validate_manager.os.path.isfile', side_effect=is_file_yml)
+        mocker.patch('demisto_sdk.commands.validate.validate_manager.find_type', return_value=FileType.INTEGRATION)
+        mocker.patch.object(ValidateManager, '_is_py_script_or_integration', return_value=False)
+
         validate_manager = ValidateManager(staged=True)
-        a = validate_manager.get_modified_and_added_files('..', 'master')
-        print(a)
+        modified_files_list, _, _, _, modified_packs = validate_manager.get_modified_and_added_files('..', 'master')
+        assert modified_files_list == {'Packs/HelloWorld/Integrations/HelloWorld.yml'}
+        assert modified_packs == {'HelloWorld'}
 
     def test_not_staged(self, mocker):
+        """
+        Given
+            - staged = False flag
+            - diff on yml file
+        When
+            - Run the validate command.
+        Then
+            - Validate that not a git diff staged command runs
+        """
         def run_command_effect(arg):
             assert 'staged' not in arg
-            return "M\tPacks/HelloWorld/Integrations/HelloWorld.py"
-
-        def is_file_yml(arg):
-            assert arg == 'Packs/HelloWorld/Integrations/HelloWorld.yml'
+            return "M\tPacks/HelloWorld/Integrations/HelloWorld.yml"
 
         mocker.patch('demisto_sdk.commands.validate.validate_manager.run_command', side_effect=run_command_effect)
-        mocker.patch('demisto_sdk.commands.validate.validate_manager.os.path.isfile', side_effect=is_file_yml)
+        mocker.patch('demisto_sdk.commands.validate.validate_manager.os.path.isfile', return_value=True)
+        mocker.patch('demisto_sdk.commands.validate.validate_manager.find_type', return_value=FileType.INTEGRATION)
+        mocker.patch.object(ValidateManager, '_is_py_script_or_integration', return_value=False)
+
         validate_manager = ValidateManager(staged=False)
-        validate_manager.get_modified_and_added_files('..', 'master')
+        modified_files_list, _, _, _, modified_packs = validate_manager.get_modified_and_added_files('..', 'master')
+        assert modified_files_list == {'Packs/HelloWorld/Integrations/HelloWorld.yml'}
+        assert modified_packs == {'HelloWorld'}
 
 
 def test_content_release_identifier_exists():